API validation

+++ api/api/Controllers/StockController.cs
@@ -27,7 +27,10 @@ namespace api.Controllers
 		[HttpGet]
 		public async Task<IActionResult> GetAll()
 		{
-			var stocks = await _stockRepo.GetAllAsync();
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            var stocks = await _stockRepo.GetAllAsync();
 
 			var stockDto = 	stocks.Select(s => s.ToStockDto());
 
@@ -38,7 +41,10 @@ namespace api.Controllers
 		[HttpGet("{id:int}")]
 		public async Task<IActionResult> GetById([FromRoute] int id)
 		{
-			var stock = await _stockRepo.GetByIdAsync(id);
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            var stock = await _stockRepo.GetByIdAsync(id);
 
 			if(stock == null)
 			{
@@ -52,7 +58,11 @@ namespace api.Controllers
 		[HttpPost]
 		public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto)
 		{
-			var stockModel = stockDto.ToStockFromCreateDTO();
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            var stockModel = stockDto.ToStockFromCreateDTO();
+
 			await _stockRepo.CreateAsync(stockModel);
 			return CreatedAtAction(nameof(GetById), new { id = stockModel.Id }, stockModel.ToStockDto());
 		}
@@ -61,7 +71,10 @@ namespace api.Controllers
 		[Route("{id:int}")]
 		public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto)
 		{
-			var stockModel = await _stockRepo.UpdateAsync(id, updateDto);
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            var stockModel = await _stockRepo.UpdateAsync(id, updateDto);
 
 			if(stockModel == null)
 			{
@@ -76,7 +89,10 @@ namespace api.Controllers
 		[Route("{id:int}")]
 		public async Task<IActionResult> Delete([FromRoute] int id)
 		{
-			var stockModel = await _stockRepo.DeleteAsync(id);
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            var stockModel = await _stockRepo.DeleteAsync(id);
 
 			if(stockModel == null)
 			{
